[smc] Initial commit with SMC with vines logic in mini-pyro format.
In particular, this includes rendering logic, the target image, and
scoring.

It isn't verified working, but it does run with N particles (and pretty
slow). Also, the weighting of the particles is likely 100% wrong haha.

Much work to be done! In particular, it may be better to write a simpler
SMC example (with this in mind).